requirements.txt
google-cloud-bigquery==3.27.0
google-cloud-secret-manager==2.21.1
requests==2.32.3
python-json-logger==2.0.7
tenacity==8.5.0
